pyproject.toml
[project]
name = "nscom01-mco2"
version = "0.1.0"
description = "Real-time Audio Streaming over IP"
requires-python = "==3.12.*"
dependencies = [
    "bitstruct>=8.22.1",
    "ffmpeg-python>=0.2.0",
    "imageio-ffmpeg>=0.6.0",
    "numpy>=2.4.4",
    "pyaudio>=0.2.14",
]
